reviewing the user/topic/babble models. adding utils and run scripts

+++ db/migrations/20201004004328-create-babble.js
@@ -29,7 +29,7 @@ module.exports = {
       },
       readTime: {
         allowNull: false,
-        type: Sequelize.NUMERIC
+        type: Sequelize.INTEGER
       },
       url: {
         type: Sequelize.STRING
